refactor: merge data-aggregation into data-loader
Move all helper functions from data-aggregation.ts into data-loader.ts
since they are only used in that file. This improves code organization
by keeping related code together and reduces unnecessary file separation.

- Moved aggregateByModel, aggregateModelBreakdowns, createModelBreakdowns,
  calculateTotals, filterByDateRange, isDuplicateEntry, markAsProcessed,
  and extractUniqueModels functions into data-loader.ts
- Made all helper functions private (no export) since they are internal
- Deleted data-aggregation.ts file
- Removed import statement from data-loader.ts

This maintains the same functionality and test coverage while
simplifying the module structure.
